web/requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pony==0.7.16
PyYAML==6.0
segno==1.5.2
Werkzeug==2.3.0
psycopg2-binary